Fix absence of B1.
The typeclass one_op (notation: 1%C) stands for the number '1', not to
be confused with the bit vector full of ones. This patch fixes a
compilation error, so it's quick & dirty. In particular, I'm not sure
it is wise to use 'bitn' to create the number '1': surely, there must
be a better way to do this directly.